fullstack/server/requirements.txt
python-dotenv
fastapi
sqlalchemy # for database
python-jose # for jwt
passlib # for password hashing